requirements.txt
Flask==2.3.2
gunicorn
undetected-chromedriver==3.5.4
Pillow==10.2.0
